refactor(api): centralize API response types and improve service imports
- Move IApiResponse type to shared types file and update all service imports
- Remove redundant send-response.util and update apiClient to use new type
- Add USER_ID cookie constant and save it during login
- Clean up imports and remove unused code in various service files
