+++ src/constants/app.constants.ts
@@ -1,4 +1,5 @@
 export enum COOKIES {
   AUTHORIZATION = "authorization",
   REFRESH_TOKEN = "refresh-token",
+  USER_ID = "user-id",
 }
