+++ src/types/api-response.ts
@@ -3,13 +3,13 @@ import { ZodIssue } from "zod";
 interface IApiResponseSuccess<T> {
   status: "success";
   message: string;
-  data: T;
+  data: T extends undefined ? never : T;
 }
 
 interface IApiResponseFailed {
   status: "failed";
   message: string;
-  code?: string;
+  code: string;
   errors?: ZodIssue[];
 }
