+++ src/types/index.ts
@@ -1,6 +1,8 @@
+export * from "./api-response";
 export * from "./auth";
 export * from "./courses";
 export * from "./otp";
 export * from "./profile";
+export * from "./quiz";
 export * from "./socket";
 export * from "./user";
